fix(audio): prevent rtkit RLIMIT_RTTIME SIGKILL during road/zone placement
Root cause: when OpenAL Soft contacts rtkit via D-Bus to request SCHED_RR
for its mixing thread, rtkit sets RLIMIT_RTTIME=200ms process-wide.  During
heavy road/zone operations (terrain flushes, mesh rebuilds stalling the main
thread), the RT mixing thread accumulates >200ms of continuous CPU time and
the kernel fires an uncatchable SIGKILL.

Fix: add `rt-prio = 0` to the [general] section of the ALSOFT config written
by AudioSystem::AudioSystem() before the first alcOpenDevice() call.  This
disables rtkit integration so the mixing thread runs as SCHED_OTHER with no
RLIMIT_RTTIME on any thread.

Also retained (belt-and-suspenders): `device = pulse` in [alsa] to route
through the PulseAudio ALSA plugin instead of PipeWire's native plugin, which
calls kill(getpid(), SIGKILL) on repeated underruns.

Additional changes in this commit:
- Remove b176ba2 FPS cap (sleep_for) and per-step timing report from main.cpp
- Remove drawScene profiling accumulators from IrrlichtRenderer (.cpp/.h)
- Remove diagnostic signal handler / std::set_terminate scaffolding from main.cpp
- Fix unsigned integer overflow UB in CitySimulation phaseOffset lambda
- Document the rtkit + PipeWire SIGKILL mechanisms in audio-system.md

+++ src/simulation/CitySimulation.cpp
@@ -1702,7 +1702,8 @@ void CitySimulation::placeRoad(int tileX, int tileZ, int earthworksCostOverride)
 
         // Helper lambda: generate staggered initial phase offset from tile coordinates.
         auto phaseOffset = [&](int cx, int cz) -> float {
-            unsigned int seed = static_cast<unsigned int>(cx * 73856093 ^ cz * 19349663);
+            unsigned int seed = (static_cast<unsigned int>(cx) * 73856093u)
+                              ^ (static_cast<unsigned int>(cz) * 19349663u);
             return (static_cast<float>(seed & 0xFFFFu) / 65535.0f) *
                    SimulationConstants::traffic_signal_phase_seconds;
         };
